don't require Boost for the actual library build, only its use
This means that we don't a have a buildreq on Boost anymore -- very
important if we ever want to get this into Qt. Just make sure that you
set BOOST_DIR and build the tests if you modify with the template parts
of the library, since otherwise they aren't even compile-tested by the
library build!