Optimization: store connection information into qobjects themselves.
Since a object knows objects it is connected to, we does not have to
iterate over all connections on its destruction to find that information.

To store the connection information, QObject::setUserData() is used.
It is noteworthy that data that has been set to an object using the function
will be automatically destroyed when the object will be destroyed. I use
the ability instead of connecting to destroyed() signal.
The downside of using setUserData is increased memory consumtion. Also,
I did not think of a way to avoid a global lock when getting a pointer
to the connection information.

The last but very important optimization is that the library keeps a
list of unused bindings (a "free list"), so when it creates a new connection
it does not iterate over all bindings to find an empty place.

Overall, the patch gives the library significant speedup. For instance,
the old version took 22 seconds to connect one signal to 100000 objects,
while the new one made it in less than 200 milliseconds. The destruction
of the objects is just slightly costly than that for the new version,
the old one can't make it in less than 40 seconds.